+++ lib/grape/exceptions/base.rb
@@ -55,7 +55,7 @@ module Grape
       end
 
       def translate_message(key, options = {})
-        translate("#{BASE_MESSAGES_KEY}.#{key}", {:default => '' }.merge(options))
+        translate("#{BASE_MESSAGES_KEY}.#{key}", { :default => '' }.merge(options))
       end
 
       def translate(key, options = {})